Move Send/RecvNotify calls outside of CopyIn/Out, due to lock order.
We were calling Send/RecvNotify inside of endpoint.Send/RecvMsg, which is
called as part of CopyIn/Out with mm.activeMu held. This leads to lock order
violation because those Send/RecvNotify needs TaskSet.mu to send a signal to a
task.

This CL changes endpoint.Send/RecvMsg to return a notification callback which
can be plumbed to the caller of CopyIn/Out and called without mm.activeMu held.

Updated the lock documentation in mm.go to order TaskSet.mu > mm.activeMu.

PiperOrigin-RevId: 451508328