Allow pod init config to be used with `runsc trace create`
As convenience allow a pod init config file to be used to create
a trace session. It's common to want to use the same configuration
for pod init and trace create, but the config file formats are
different. This change allows both formats to be used with
`runsc trace create`.

Updates #4805

PiperOrigin-RevId: 452142362